spec/html5lib/test.dat

#data
<!doctype html><html><body>hello<!--X--></body></html>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <body>
|     "hello"
|     <!-- X -->

#data
<!doctype html><html><head><title>hi</title></head><body>hello<!--X--></body></html>
#errors
#document
| <!DOCTYPE html>
| <html>
|   <head>
|     <title>
|       "hi"
|   <body>
|     "hello"
|     <!-- X -->
